FEAT: Completed login flow with automatic redirection based on auth state (presence of a valid token
