+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^7.1.2",
     "axios": "^0.19.0",
     "bulma": "^0.8.0",
+    "history": "^4.10.1",
     "jwt-decode": "^2.2.0",
     "node-sass": "^4.13.0",
     "react": "^16.12.0",
